+++ SourceCode/Database/dbo/Procedures/AddGeneratedHubWaybills.sql
@@ -1,4 +1,4 @@
-﻿CREATE PROCEDURE [dbo].[AddGeneratedModuleWaybills]
+﻿CREATE PROCEDURE [dbo].[AddGeneratedHubWaybills]
 	@StationCustomerId INT
 AS
 -- The columns in this stored procedures must match table StationCustomerWaybill
@@ -15,7 +15,7 @@ BEGIN
 			StationCustomerWaybill AS SCW ON ME.StationCustomerCargoId = SCW.StationCustomerCargoId AND 
 			OTHER.StationCustomerCargoId = SCW.OtherStationCustomerCargoId
 		WHERE
-			ME.NHMCode > 0 AND ME.StationCustomerId = @StationCustomerId AND SCW.Id IS NOT NULL
+			ME.NHMCode = 0 AND  ME.StationCustomerId = @StationCustomerId AND SCW.Id IS NOT NULL
 	)
 	UPDATE StationCustomerWaybill 
 		SET 
@@ -30,7 +30,7 @@ BEGIN
 	SELECT
 		ME.StationCustomerId AS [StationCustomerId],
 		ME.StationCustomerCargoId AS [StationCustomerCargoId],
-		OTHER.StationCustomerCargoId AS [OtherStationCustomerCargoId],
+		NULL AS [OtherStationCustomerCargoId],
 		NULL AS [OtherExternalCustomerCargoId],
 		OTHER.RegionId AS [OtherRegionId],
 		ME.OperatingDayId AS OperatingDayId,
@@ -46,15 +46,16 @@ BEGIN
 		ModuleCustomerCargo AS OTHER ON ME.CargoId = OTHER.CargoId LEFT JOIN
 		StationCustomerWaybill AS SCW ON ME.StationCustomerCargoId = SCW.StationCustomerCargoId AND OTHER.StationCustomerCargoId = SCW.OtherStationCustomerCargoId
 	WHERE
-		ME.NHMCode > 0 AND
-		ME.StationCustomerId = @StationCustomerId
+		OTHER.IsCargoHub <> 0
+		AND ME.NHMCode = 0 
+		AND ME.StationCustomerId = @StationCustomerId
 		AND ME.StationCustomerId <> OTHER.StationCustomerId
 		AND ME.IsSupply <> OTHER.IsSupply AND ME.IsSupply = 0
 		--AND ME.QuantityUnitId = OTHER.QuantityUnitId -- Is not used, because it causes too few matches.
 		AND ME.StationId <> OTHER.StationId
 		AND ME.MainTheme = OTHER.MainTheme
 		AND (ME.ScaleId = OTHER.ScaleId OR ME.ScaleId = 0 OR OTHER.ScaleId = 0)
-		AND (ME.CountryId = OTHER.CountryId OR ME.IsInternational <> 0 AND OTHER.IsInternational <>0) 
+		AND (ME.CountryId = OTHER.CountryId) 
 		AND (ME.FromYear IS NULL OR OTHER.UptoYear IS NULL OR ME.FromYear <= OTHER.UptoYear )
 		AND (ME.UptoYear IS NULL OR OTHER.FromYear IS NULL OR ME.UptoYear >= OTHER.FromYear )
 		AND SCW.Id IS NULL